pyproject.toml
[tool.poetry]
name = "django-public-admin"
version = "0.0.6"
description = "A public read-only version of the Django Admin"
authors = ["Eduardo Cuducos <cuducos@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/cuducos/django-public-admin"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Software Development",
  "Framework :: Django",
  "Framework :: Django :: 3.2",
  "Framework :: Django :: 4.2",
  "Framework :: Django :: 5.0",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
packages = [
    {include = "public_admin"}
]

[tool.poetry.dependencies]
python = "^3.9"
django = ">=3.2"

[tool.poetry.dev-dependencies]
pytest-ruff = "^0.3.1"
readthedocs-sphinx-ext = "^2.2.5"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"
tox = "^4.14.2"

[tool.pytest.ini_options]
addopts = "--ruff --ruff-format"

[tool.ruff.lint]
select = ["E", "F", "I"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
